[Issue-3423] refactor: minor refactor
